Reorganize wallets/
* Move passphrase hooks into own file
* Rename to draggable.js to dnd.js
* Create wallets/lib/crypto.js
* Move wallets/client/context stuff into ../hooks

The distinction between context and hooks did not make sense.

* Fix circular imports within wallets/client/hooks
