+++ wallets/client/hooks/diagnostics.js
@@ -1,8 +1,8 @@
+import { useCallback } from 'react'
+import { useMutation } from '@apollo/client'
 import { useMe } from '@/components/me'
 import { useToast } from '@/components/toast'
 import { SET_DIAGNOSTICS } from '@/fragments/users'
-import { useMutation } from '@apollo/client'
-import { useCallback } from 'react'
 
 export function useDiagnostics () {
   const { me, refreshMe } = useMe()
