+++ wallets/client/hooks/index.js
@@ -1,3 +1,4 @@
+export * from './global'
 export * from './payment'
 export * from './image'
 export * from './indicator'
@@ -7,3 +8,5 @@ export * from './crypto'
 export * from './query'
 export * from './logger'
 export * from './diagnostics'
+export * from './passphrase'
+export * from './dnd'
