+++ wallets/client/hooks/prompt.js
@@ -1,13 +1,13 @@
 import { useCallback } from 'react'
 import { boolean, object } from 'yup'
 import { Button } from 'react-bootstrap'
+import Link from 'next/link'
+import { useMutation } from '@apollo/client'
+import styles from '@/styles/wallet.module.css'
+import { HIDE_WALLET_RECV_PROMPT_MUTATION } from '@/fragments/users'
 import { Form, ClientInput, SubmitButton, Checkbox } from '@/components/form'
 import { useMe } from '@/components/me'
 import { useShowModal } from '@/components/modal'
-import Link from 'next/link'
-import styles from '@/styles/wallet.module.css'
-import { useMutation } from '@apollo/client'
-import { HIDE_WALLET_RECV_PROMPT_MUTATION } from '@/fragments/users'
 import { useToast } from '@/components/toast'
 import { useLightningAddressUpsert } from '@/wallets/client/hooks/query'
 import { protocolClientSchema } from '@/wallets/lib/util'
