+++ wallets/lib/crypto.js
@@ -1,0 +1,83 @@
+import bip39Words from '@/lib/bip39-words'
+
+export async function deriveKey (passphrase, salt) {
+  const enc = new TextEncoder()
+
+  const keyMaterial = await window.crypto.subtle.importKey(
+    'raw',
+    enc.encode(passphrase),
+    { name: 'PBKDF2' },
+    false,
+    ['deriveKey']
+  )
+
+  const key = await window.crypto.subtle.deriveKey(
+    {
+      name: 'PBKDF2',
+      salt: enc.encode(salt),
+      // 600,000 iterations is recommended by OWASP
+      // see https://cheatsheetseries.owasp.org/cheatsheets/Password_Storage_Cheat_Sheet.html#pbkdf2
+      iterations: 600_000,
+      hash: 'SHA-256'
+    },
+    keyMaterial,
+    { name: 'AES-GCM', length: 256 },
+    true,
+    ['encrypt', 'decrypt']
+  )
+
+  const rawKey = await window.crypto.subtle.exportKey('raw', key)
+  const hash = Buffer.from(await window.crypto.subtle.digest('SHA-256', rawKey)).toString('hex')
+  const unextractableKey = await window.crypto.subtle.importKey(
+    'raw',
+    rawKey,
+    { name: 'AES-GCM' },
+    false,
+    ['encrypt', 'decrypt']
+  )
+
+  return {
+    key: unextractableKey,
+    hash
+  }
+}
+
+export async function encrypt ({ key, hash }, value) {
+  // random IVs are _really_ important in GCM: reusing the IV once can lead to catastrophic failure
+  // see https://crypto.stackexchange.com/questions/26790/how-bad-it-is-using-the-same-iv-twice-with-aes-gcm
+  // 12 bytes (96 bits) is the recommended IV size for AES-GCM
+  const iv = window.crypto.getRandomValues(new Uint8Array(12))
+  const encoded = new TextEncoder().encode(JSON.stringify(value))
+  const encrypted = await window.crypto.subtle.encrypt(
+    {
+      name: 'AES-GCM',
+      iv
+    },
+    key,
+    encoded
+  )
+  return {
+    keyHash: hash,
+    iv: Buffer.from(iv).toString('hex'),
+    value: Buffer.from(encrypted).toString('hex')
+  }
+}
+
+export async function decrypt (key, { iv, value }) {
+  const decrypted = await window.crypto.subtle.decrypt(
+    {
+      name: 'AES-GCM',
+      iv: Buffer.from(iv, 'hex')
+    },
+    key,
+    Buffer.from(value, 'hex')
+  )
+  const decoded = new TextDecoder().decode(decrypted)
+  return JSON.parse(decoded)
+}
+
+export function generateRandomPassphrase () {
+  const rand = new Uint32Array(12)
+  window.crypto.getRandomValues(rand)
+  return Array.from(rand).map(i => bip39Words[i % bip39Words.length]).join(' ')
+}
